news_crawler/requirements.txt
newspaper3k
nltk
sumy